[IMP] stock_reception_screen: keep last lot id used
When a stock move is received in multiple steps, the lot id has to be
entered everytime, which is a lost of time.
Here we keep the last lot_id used for a move and it is set as default
for the next move line to be received.
It can always be changed if needed.

And fix the reset button.